+++ test/ui/commands/test_list.py
@@ -1,3 +1,6 @@
+import pytest
+
+from beets.exceptions import UserError
 from beets.test import _common
 from beets.test.helper import BeetsTestCase, IOMixin
 
@@ -5,10 +8,11 @@ from beets.test.helper import BeetsTestCase, IOMixin
 class ListTest(IOMixin, BeetsTestCase):
     def setUp(self):
         super().setUp()
-        self.item = _common.item()
-        self.item.path = "xxx/yyy"
+        self.item = _common.item(path="xxx/yyy", flex=1)
         self.lib.add(self.item)
         self.lib.add_album([self.item])
+        self.another_item = _common.item(path="another/path", flex=2)
+        self.lib.add(self.another_item)
 
     def test_list_outputs_item(self):
         stdout = self.run_with_output("list")
@@ -25,7 +29,7 @@ class ListTest(IOMixin, BeetsTestCase):
 
     def test_list_item_path(self):
         stdout = self.run_with_output("list", "-f", "$path")
-        assert stdout.strip() == str(self.lib_path / "xxx/yyy")
+        assert stdout.strip().splitlines()[0] == str(self.lib_path / "xxx/yyy")
 
     def test_list_album_outputs_something(self):
         stdout = self.run_with_output("list", "-a")
@@ -55,9 +59,21 @@ class ListTest(IOMixin, BeetsTestCase):
 
     def test_list_item_format_multiple(self):
         stdout = self.run_with_output("list", "-f", "$artist - $album - $year")
-        assert "the artist - the album - 0001" == stdout.strip()
+        assert "the artist - the album - 0001" == stdout.strip().splitlines()[0]
 
     def test_list_album_format(self):
         stdout = self.run_with_output("list", "-a", "-f", "$genres")
         assert "the genre" in stdout
         assert "the album" not in stdout
+
+    def test_limit_query_results(self):
+        args = "list", "-p"
+
+        stdout = self.run_with_output(*args).strip()
+        assert len(stdout.splitlines()) == 2
+
+        stdout = self.run_with_output(*args, "-l", "1").strip()
+        assert len(stdout.splitlines()) == 1
+
+        with pytest.raises(UserError, match="must be a non-negative integer"):
+            self.run_with_output(*args, "-l", "-1")